requirements.txt
twisted
pyserial